Use predefined seed for randomizations

+++ Assets/Scripts/Database/Models/DatabaseModels.cs
@@ -63,7 +63,7 @@ namespace Simulator.Database
         public float? Fog { get; set; }
         public float? Wetness { get; set; }
         public float? Cloudiness { get; set; }
-        public long? Seed { get; set; }
+        public int? Seed { get; set; }
         public bool? UseTraffic { get; set; }
         public bool? UsePedestrians { get; set; }
         public bool? UseBicyclists { get; set; }
